Restructure sync logic into folders, to accomodate for testing
