+++ src/index.ts
@@ -2,8 +2,8 @@ import { Plugin } from 'obsidian';
 import { get } from 'svelte/store';
 
 import FileManager from './fileManager';
-import SyncHighlights from './syncHighlights';
-import SyncKindleClippings from './syncKindleClippings';
+import SyncHighlights from './sync/syncHighlights';
+import SyncKindleClippings from './sync/syncKindleClippings';
 import SyncModal from './components/syncModal';
 import { SettingsTab } from './settingsTab';
 import { StatusBar } from './components/statusBar';
